rename: file names to match hypercore action name
