+++ packages/hyperliquid-composer/src/commands/index.ts
@@ -1,4 +1,4 @@
 export * from './set-block'
 export * from './register-token'
 export * from './core-spot-deployment'
-export * from './deploy-spot'
+export * from './spot-deploy'
