+++ packages/hyperliquid-composer/src/commands/spot-deploy.ts
